pyproject.toml
[project]
name = "project2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "numpy>=2.2.6",
    "opencv-contrib-python>=4.12.0.88",
    "opencv-python>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=11.3.0",
    "streamlit>=1.49.1",
    "tensorflow>=2.20.0",
]
